rename: utils/expt_utils -> utils/diffraction
