+++ pyxem/utils/polar_transform_utils.py
@@ -17,7 +17,7 @@
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
-from pyxem.utils.expt_utils import find_beam_center_blur
+from pyxem.utils.diffraction import find_beam_center_blur
 from scipy import ndimage
 from pyxem.utils.cuda_utils import get_array_module
